Fix invalid handle accessors between raw and other types
