+++ cscore/src/main/native/cpp/ConfigurableSourceImpl.cpp
@@ -104,3 +104,139 @@ void ConfigurableSourceImpl::SetEnumPropertyChoices(
                                   prop->name, property, CS_PROP_ENUM,
                                   prop->value, {});
 }
+
+namespace cs {
+static constexpr unsigned SourceMask = CS_SOURCE_CV | CS_SOURCE_RAW;
+
+void NotifySourceError(CS_Source source, std::string_view msg,
+                       CS_Status* status) {
+  auto data = Instance::GetInstance().GetSource(source);
+  if (!data || (data->kind & SourceMask) == 0) {
+    *status = CS_INVALID_HANDLE;
+    return;
+  }
+  static_cast<ConfigurableSourceImpl&>(*data->source).NotifyError(msg);
+}
+
+void SetSourceConnected(CS_Source source, bool connected, CS_Status* status) {
+  auto data = Instance::GetInstance().GetSource(source);
+  if (!data || (data->kind & SourceMask) == 0) {
+    *status = CS_INVALID_HANDLE;
+    return;
+  }
+  static_cast<ConfigurableSourceImpl&>(*data->source).SetConnected(connected);
+}
+
+void SetSourceDescription(CS_Source source, std::string_view description,
+                          CS_Status* status) {
+  auto data = Instance::GetInstance().GetSource(source);
+  if (!data || (data->kind & SourceMask) == 0) {
+    *status = CS_INVALID_HANDLE;
+    return;
+  }
+  static_cast<ConfigurableSourceImpl&>(*data->source).SetDescription(description);
+}
+
+CS_Property CreateSourceProperty(CS_Source source, std::string_view name,
+                                 CS_PropertyKind kind, int minimum, int maximum,
+                                 int step, int defaultValue, int value,
+                                 CS_Status* status) {
+  auto data = Instance::GetInstance().GetSource(source);
+  if (!data || (data->kind & SourceMask) == 0) {
+    *status = CS_INVALID_HANDLE;
+    return -1;
+  }
+  int property = static_cast<ConfigurableSourceImpl&>(*data->source)
+                     .CreateProperty(name, kind, minimum, maximum, step,
+                                     defaultValue, value);
+  return Handle{source, property, Handle::kProperty};
+}
+
+CS_Property CreateSourcePropertyCallback(
+    CS_Source source, std::string_view name, CS_PropertyKind kind, int minimum,
+    int maximum, int step, int defaultValue, int value,
+    std::function<void(CS_Property property)> onChange, CS_Status* status) {
+  auto data = Instance::GetInstance().GetSource(source);
+  if (!data || (data->kind & SourceMask) == 0) {
+    *status = CS_INVALID_HANDLE;
+    return -1;
+  }
+  int property = static_cast<ConfigurableSourceImpl&>(*data->source)
+                     .CreateProperty(name, kind, minimum, maximum, step,
+                                     defaultValue, value, onChange);
+  return Handle{source, property, Handle::kProperty};
+}
+
+void SetSourceEnumPropertyChoices(CS_Source source, CS_Property property,
+                                  std::span<const std::string> choices,
+                                  CS_Status* status) {
+  auto data = Instance::GetInstance().GetSource(source);
+  if (!data || (data->kind & SourceMask) == 0) {
+    *status = CS_INVALID_HANDLE;
+    return;
+  }
+
+  // Get property index; also validate the source owns this property
+  Handle handle{property};
+  int i = handle.GetParentIndex();
+  if (i < 0) {
+    *status = CS_INVALID_HANDLE;
+    return;
+  }
+  auto data2 = Instance::GetInstance().GetSource(Handle{i, Handle::kSource});
+  if (!data2 || data->source.get() != data2->source.get()) {
+    *status = CS_INVALID_HANDLE;
+    return;
+  }
+  int propertyIndex = handle.GetIndex();
+  static_cast<ConfigurableSourceImpl&>(*data->source)
+      .SetEnumPropertyChoices(propertyIndex, choices, status);
+}
+
+}  // namespace cs
+
+extern "C" {
+void CS_NotifySourceError(CS_Source source, const char* msg,
+                          CS_Status* status) {
+  return cs::NotifySourceError(source, msg, status);
+}
+
+void CS_SetSourceConnected(CS_Source source, CS_Bool connected,
+                           CS_Status* status) {
+  return cs::SetSourceConnected(source, connected, status);
+}
+
+void CS_SetSourceDescription(CS_Source source, const char* description,
+                             CS_Status* status) {
+  return cs::SetSourceDescription(source, description, status);
+}
+
+CS_Property CS_CreateSourceProperty(CS_Source source, const char* name,
+                                    enum CS_PropertyKind kind, int minimum,
+                                    int maximum, int step, int defaultValue,
+                                    int value, CS_Status* status) {
+  return cs::CreateSourceProperty(source, name, kind, minimum, maximum, step,
+                                  defaultValue, value, status);
+}
+
+CS_Property CS_CreateSourcePropertyCallback(
+    CS_Source source, const char* name, enum CS_PropertyKind kind, int minimum,
+    int maximum, int step, int defaultValue, int value, void* data,
+    void (*onChange)(void* data, CS_Property property), CS_Status* status) {
+  return cs::CreateSourcePropertyCallback(
+      source, name, kind, minimum, maximum, step, defaultValue, value,
+      [=](CS_Property property) { onChange(data, property); }, status);
+}
+
+void CS_SetSourceEnumPropertyChoices(CS_Source source, CS_Property property,
+                                     const char** choices, int count,
+                                     CS_Status* status) {
+  wpi::SmallVector<std::string, 8> vec;
+  vec.reserve(count);
+  for (int i = 0; i < count; ++i) {
+    vec.push_back(choices[i]);
+  }
+  return cs::SetSourceEnumPropertyChoices(source, property, vec, status);
+}
+
+}  // extern "C"
